+++ crime-hardship/src/main/java/uk/gov/justice/laa/crime/hardship/controller/HardshipController.java
@@ -20,7 +20,7 @@ import uk.gov.justice.laa.crime.hardship.model.*;
 import uk.gov.justice.laa.crime.hardship.service.CrimeMeansAssessmentService;
 import uk.gov.justice.laa.crime.hardship.service.HardshipCalculationService;
 import uk.gov.justice.laa.crime.hardship.service.HardshipService;
-import uk.gov.justice.laa.crime.hardship.service.HardshipValidationService;
+import uk.gov.justice.laa.crime.hardship.validation.HardshipValidationService;
 import uk.gov.justice.laa.crime.hardship.staticdata.enums.HardshipReviewDetailType;
 import uk.gov.justice.laa.crime.hardship.staticdata.enums.RequestType;
