+++ crime-hardship/src/test/java/uk/gov/justice/laa/crime/hardship/service/HardshipValidationServiceTest.java
@@ -11,6 +11,7 @@ import uk.gov.justice.laa.crime.hardship.model.HardshipMetadata;
 import uk.gov.justice.laa.crime.hardship.model.HardshipReview;
 import uk.gov.justice.laa.crime.hardship.model.SolicitorCosts;
 import uk.gov.justice.laa.crime.hardship.staticdata.enums.*;
+import uk.gov.justice.laa.crime.hardship.validation.HardshipValidationService;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
